Refactor pymantic.parsers to not be one megamodule
