+++ pymantic/parsers/__init__.py
@@ -1,0 +1,8 @@
+__all__ = ['ntriples_parser', 'nquads_parser', 'turtle_parser', 'jsonld_parser']
+
+from .lepl import (
+    ntriples_parser,
+    nquads_parser,
+    turtle_parser,
+    jsonld_parser,
+)
